Use installExtension in showTutorial
