+++ src/commands/trialApp/showTutorial.ts
@@ -3,16 +3,13 @@
 *  Licensed under the MIT License. See License.txt in the project root for license information.
 *--------------------------------------------------------------------------------------------*/
 
-import { commands, extensions, Uri } from 'vscode';
+import { commands, Uri } from 'vscode';
 import { AzExtTreeItem, IActionContext } from 'vscode-azureextensionui';
 import { TrialAppTreeItem } from '../../explorer/trialApp/TrialAppTreeItem';
-import { delay } from '../../utils/delay';
+import { installExtension } from '../../utils/installExtension';
 import { ext } from './../../extensionVariables';
 
 export async function showTutorial(context: IActionContext, node?: TrialAppTreeItem): Promise<void> {
-
-    const timeoutInSeconds: number = 60;
-    const maxTime: number = Date.now() + timeoutInSeconds * 1000;
     const extensionId: string = 'redhat.vscode-didact';
 
     if (!node) {
@@ -24,26 +21,8 @@ export async function showTutorial(context: IActionContext, node?: TrialAppTreeI
         });
     }
 
-    if (extensions.getExtension(extensionId)) {
-        showDidact();
-    } else {
-        const commandToRun: string = 'extension.open';
-        commands.executeCommand(commandToRun, extensionId);
-
-        while (Date.now() < maxTime) {
-
-            if (extensions.getExtension(extensionId)) {
-                showDidact();
-                break;
-            } else {
-                await delay(5000);
-            }
-        }
+    if (await installExtension(extensionId)) {
+        const tutorialUri: Uri = Uri.file(ext.context.asAbsolutePath('resources/TrialApp.didact.md'));
+        commands.executeCommand('vscode.didact.startDidact', tutorialUri);
     }
-}
-
-function showDidact(): void {
-
-    const tutorialUri: Uri = Uri.file(ext.context.asAbsolutePath('resources/TrialApp.didact.md'));
-    commands.executeCommand('vscode.didact.startDidact', tutorialUri);
 }
